Rework room shaders
A room can now have multiple shaders that are sorted and applied in
order over the screen (by passing them a clear texture and then
drawing it on the screen over the target texture).

This also makes it possible to separate the light as a different
shader. The screen light has been replaced with a simpler spherical
light.