Implement overlay drawing
This is to be used for minigames and such, which should be drawn over
the already shaded game, without shaders. This also introduces a
concept of layers into the drawable, although it doesn't do much
except not drawing the layered drawables with all the others and only
drawing them when requested.