Reorganize the walking part a lot
Things not live in different files for that most part, and not in one
big file. This creates some difficulties with circular dependencies,
which are avoided with module lazy loading. The rooms module stores
things more-or-less related to rooms, the entities module stores
things related to entities, and other modules are dedicated to their
components and systems, which are then required and loaded from the
rooms and entities modules.