Allow callbacks to specify functions in modules, not just names
This allows to separate each callback to their module fully, with that
module exporting a field named interaction_callbacks or
activatable_callbacks. The functions in these fields are then used for
those callbacks. Also, all the callback-related data is now grouped
into one table in TOML, and not separated through the component, the
args became callback.args, and activatable_callback gained args too.