Implement allocation arguments with patched middleclass, line_source
With a patched version of middleclass arguments like character and
name can be provided before the constructor is called, so the method
update_for_character is not needed anymore.

Line source is the source from which lines take their next line when
they want to make the next one. This was added for mod support.