Modify the sound player, first puzzle room and add wind
The sound player is modified to allow looping and positional sound, to
allow playing the wind sound effect as if it's somewhere above. Also,
the activatable callback is made optional, so that the wind sound can
play from the start.

The jingle player is removed from the first puzzle room on day 4, so
it won't play in case something goes wrong with state variables (or if
the room was loaded manually).

For systems, a onBeforeResetEngine callback is now called if it
exists. It's called before a room is loaded, so the systems may stop
doing whatever they do (like playing sound, in case of sound playing
system), as SFML isn't directly tied to lua.

Finally, a wind sound effect is added to the walkway room, this sound
effect is played that is supposed to be coming from somewhere above
the player, the sound effect is looped.