Implement some editing capabilities, for Transformable and Collider
This commit introduces a base for the editor system of components,
including serialization back to files in the source code in
run-time. A read-only editors are implemented for Transformable and
Name, and for Collider an editor that actually does things is
implemented (it also updates transformable in TOML source when
needed).

As of now, new components can't be added, and neither can new
entities, but the existing components can be edited. This includes
checks for defaults, which are not written to the file if they are not
overriden in the editor. Otherwise, if the default value is changed
the key-value pair will be added to the TOML code.

The serialization-back system itself is implemented by finding the
relevant place in the TOML file and replacing just this place with
manual string manipulation on the C++ side. This is done so as to save
the formatting and such of the original file, as it seems that no TOML
libraries are able to do so and always re-format and re-arrange the
file. As such, a system to only modify the data in relevant parts was
made.

As for prefabs, currently editing anything using non-overriden prefab
value will modify the prefab itself. This will probably have to be
changed in the future somehow.