Implenent auto-loading/listing all components from the rockspec
Now the components are all loaded automatically without the need to
write them out every time. They are also sorted based on
system_run_priority beforehand because some systems need to run first
and others later.

With this change, it should be easier to implement adding components
to entities with the debug menu, as they should all be available in
Component.all (as provided by lovetoys).

"Oh cirno-chan, I wrote this code thinking of (You) :cirnoBaka:"