Implement adding components and selecting entities w/o collider
Adding components adds the table header to the file, then re-parses it
and uses that new position to actually insert the fields into it
afterwards.

Creating components is only available for components that expose
default data for now, because adding invalid default data could break
everything, since the game is not paused when editing and
such. Perpahs a better solution could be developed but for now this
should be enough.

There's a button to add a new entity but it doesn't do anything yet.

Also missing is the ability to remove components from entities.