Implement part of the activatable editor + allow deleting nodes
TOML now can recognize and delete nodes, and serialization will check
if the default value is used and then delete the node instead. Also,
now the list of the names is passed, and values are passed separately,
so if the names is in the list but in the value table it is nil, this
means that the data was deleted and should be removed from TOML.

The interactable editor includes many things that should later be put
into separate editor helper functions, such as the editor for
callbacks and later for other things that are somewhat universal.