Create/Delete components for entities with prefabs as needed
In particular, if a component doesn't override anything in the prefab,
then there's no need to create it. If it specifies the same values as
the prefab, then these parts need not be output. If no parts need to
be output, then the component entry can be deleted from the room file,
as everything is already set in the prefab.

Additionally, now when creating an entity and not overriding anything
at all, an entry is still created, where the entry only specifies the
prefab and no components.