Add an editor for the passage comp + loading entities from room
There is now a method to preload entities in a room to check something
with them without actually putting them into the world. This is
probably pretty expensive, so it's for debug only. The passage
component editor uses this method to verify the validity of the
passage before allowing to save it to a file.