Implement asset editing, addition and deletion
The debug menu now has two tabs, one for the room stuff and one for
the assets. The assets can be added, edited and removed. When editing
or removing assets, the old ones are replaced with the new ones in
runtime, putting a placeholder (at least for textures) when necessary.

The toml saving for assets works on per-asset base, it edits them one
by one, similiar to components. New ones are added at the end,
removing them also removes the newline afterwards so it looks like
there never was anything there.

The asset editing menu can search by asset name and shows a file
listing dialogue when selecting a path for a new asset, which also
allows to search in the directory by filename.

There perhaps needs to be some generic asset selection menu with
search built-in, as selecting them with a dropdown box has proven to
be inconvenient in component editors.