Fix an inconsistent bug with new component placement
The code tried to get the "self" node on every entity, even if one
did not exist (such a node would have a path like "entities.entity
name", with no component). If such a node does not exist in a file,
but its source is retreived from code, it will point to some other
node that actually exists which DOES NOT WORK and breaks things.

So, only create the "self" node in case there are no other nodes but
the entity exists in the file: therefore it is guaranteed that the
node will be pointing to an entity node with no components.