Initial replacement of SFML audio with SDL_Mixer
SDL mixer can, unlike SFML, link to the audio libraries
statically. Plus there is an option to disable the unneeded audio
formats, like FLAC. Therefore, now only WAV and OGG are supported, and
the game only uses OGG files anyway.

The interface is wrapped to be almost exactly the same as SFML.