Implement running mods from cmdline and straight into walking mode
New dependency is introduced, the "args" library, used for parsing
commandline arguments. As of now, this only adds one argument (besides
help): --load-mod, which allows to boot straight into a mod, instead
of running the main game files. This includes starting in walking mode
if needed and never using the terminal mode if so desired, or
switching to walking mode at some point.