Implement additional custom things for notes
Like a custom close text and callbacks when the particular note page
is closed, so you only go back to the terminal after reading the
SECOND page of the cave note.

Also, some fixes for rooms which don't have the "player" stuff. Should
maybe modify the player things to be overridable by the room or
something, should be useful for mods too.

No story for evening of day 4 in this commit, but it's set up, so that
the talk about the contents of the note may start after perhaps some
description.